+++ dali/placer/global_placer/GPSimPL/boxbin.cc
@@ -93,7 +93,7 @@ void BoxBin::update_cell_area_white_space(std::vector<std::vector<GridBin> > &gr
 }
 
 void BoxBin::UpdateCellAreaWhiteSpaceFillingRate(
-    std::vector<std::vector<unsigned long int>> &grid_bin_white_space_LUT,
+    std::vector<std::vector<unsigned long long>> &grid_bin_white_space_LUT,
     std::vector<std::vector<GridBin>> &grid_bin_matrix
 ) {
   if (ll_index.x == 0) {
@@ -339,12 +339,12 @@ bool BoxBin::write_cell_in_box(
   return true;
 }
 
-unsigned long int BoxBin::white_space_LUT(
-    std::vector<std::vector<unsigned long int>> &grid_bin_white_space_LUT,
+unsigned long long BoxBin::white_space_LUT(
+    std::vector<std::vector<unsigned long long>> &grid_bin_white_space_LUT,
     GridBinIndex &ll,
     GridBinIndex &ur
 ) {
-  unsigned long int white_space;
+  unsigned long long white_space;
   if (ll.x == 0) {
     if (ll.y == 0) {
       white_space = grid_bin_white_space_LUT[ur.x][ur.y];
@@ -366,9 +366,9 @@ unsigned long int BoxBin::white_space_LUT(
   return white_space;
 }
 
-bool BoxBin::update_cut_index_white_space(std::vector<std::vector<unsigned long int>> &grid_bin_white_space_LUT) {
+bool BoxBin::update_cut_index_white_space(std::vector<std::vector<unsigned long long>> &grid_bin_white_space_LUT) {
   double error, minimum_error = 1;
-  unsigned long int white_space_low;
+  unsigned long long white_space_low;
   int index_give_minimum_error;
   if (cut_direction_x) {
     if (ll_index.y == ur_index.y) return false;
@@ -421,11 +421,11 @@ bool BoxBin::update_cut_index_white_space(std::vector<std::vector<unsigned long 
 
 bool BoxBin::update_cut_point_cell_list_low_high(
     std::vector<Block> &Nodelist,
-    unsigned long int &box1_total_white_space,
-    unsigned long int &box2_total_white_space
+    unsigned long long &box1_total_white_space,
+    unsigned long long &box2_total_white_space
 ) {
   // this member function will be called only when two white spaces are not different from each other for several magnitudes
-  unsigned long int cell_area_low = 0;
+  unsigned long long cell_area_low = 0;
   double cut_line_low, cut_line_high;
   double cut_line = 0;
   double ratio =
@@ -518,7 +518,7 @@ bool BoxBin::update_cut_point_cell_list_low_high_leaf(
     int &cut_line_w,
     int ave_blk_height
 ) {
-  unsigned long int cell_area_low = 0;
+  unsigned long long cell_area_low = 0;
   double cut_line = 0;
   // the above three cut-lines are for cells, thus they needs to be double
   double ratio = 2.0;
@@ -571,7 +571,7 @@ bool BoxBin::update_cut_point_cell_list_low_high_leaf(
     }
 
     /* and then, find the index of cell, the total cell area below which is closest to one half of the total cell area */
-    unsigned long int tmp_tot_cell_area_low = 0;
+    unsigned long long tmp_tot_cell_area_low = 0;
     int index_tot_cell_low_closest_to_half = 0;
     for (size_t i = 0; i < cell_list.size(); i++) {
       node = &Nodelist[cell_list[i]];
@@ -636,7 +636,7 @@ bool BoxBin::update_cut_point_cell_list_low_high_leaf(
       cell_list[i] = tmp_cell_id;
     }
     /* second, find the index of cell, the total cell area below which is closest to one half of the total cell area */
-    unsigned long int tmp_tot_cell_area_low = 0;
+    unsigned long long tmp_tot_cell_area_low = 0;
     int index_tot_cell_low_closest_to_half = 0;
     for (size_t i = 0; i < cell_list.size(); i++) {
       node = &Nodelist[cell_list[i]];
